shared library, rename dispatch -> logistics
