+++ src/redux/reducers.js
@@ -24,7 +24,7 @@ import { appReducer } from './App'
 import accountReducer from './Account/reducers'
 import restaurantReducer from './Restaurant/reducers'
 import checkoutReducer from './Checkout/reducers'
-import { reducer as coreDispatchReducer } from '../coopcycle-frontend-js/dispatch/redux'
+import { reducer as coreDispatchReducer } from '../coopcycle-frontend-js/logistics/redux'
 import appDispatchReducer from './Dispatch/reducers'
 import storeReducer from './Store/reducers'
 import { createTaskItemsTransform } from './util'
